requirements.txt
numpy>=1.26.1
pyglet>=2.0.9
pyobjc-framework-Quartz>=11.0